+++ package.json
@@ -35,6 +35,7 @@
     "@hedviginsurance/cli": "^1.4.1",
     "@hedviginsurance/web-survival-kit": "^1.4.0",
     "@types/koa": "^2.0.46",
+    "@types/koa-basic-auth": "^2.0.3",
     "@types/koa-router": "^7.0.32",
     "@types/ramda": "^0.25.38",
     "@types/react": "^16.4.14",
@@ -48,6 +49,7 @@
     "emotion-server": "^9.2.10",
     "formik": "^1.3.0",
     "koa": "^2.5.3",
+    "koa-basic-auth": "^3.0.0",
     "koa-router": "^7.4.0",
     "ramda": "^0.25.0",
     "react": "^16.5.2",
