+++ src/serverEntry.tsx
@@ -1,57 +1,9 @@
 import 'source-map-support/register'
-import {
-  // tslint:disable-line ordered-imports
-  createKoaServer,
-  getScriptLocation,
-} from '@hedviginsurance/web-survival-kit'
-import { renderStylesToString } from 'emotion-server'
-import * as Koa from 'koa'
-import * as path from 'path'
-import * as React from 'react'
-import { renderToString } from 'react-dom/server'
-import { StaticRouter, StaticRouterContext } from 'react-router'
-
-import { App } from './App'
+import { createKoaServer } from '@hedviginsurance/web-survival-kit' // tslint:disable-line ordered-imports
 import { reactPageRoutes } from './routes'
+import { getPage } from './server/page'
+import { notNullable } from './utils/nullables'
 
-const scriptLocation = getScriptLocation({
-  statsLocation: path.resolve(__dirname, 'assets'),
-  webpackPublicPath: process.env.WEBPACK_PUBLIC_PATH || '',
-})
-const template = (body: string) => `
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Example web app</title>
-</head>
-<body>
-  <div id="react-root">${body}</div>
-  
-  <script src="${scriptLocation}"></script>
-</body>
-</html>
-`
-
-const getPage: Koa.Middleware = async (ctx) => {
-  const routerContext: StaticRouterContext = {}
-  const reactBody = renderStylesToString(
-    renderToString(
-      <StaticRouter location={ctx.request.originalUrl} context={routerContext}>
-        <App />
-      </StaticRouter>,
-    ),
-  )
-
-  if (routerContext.url) {
-    ctx.redirect(routerContext.url)
-    return
-  }
-
-  ctx.body = template(reactBody)
-}
 const getPort = () => (process.env.PORT ? Number(process.env.PORT) : 8080)
 
 console.log(`Booting server on ${getPort()} 👢`) // tslint:disable-line no-console
@@ -60,6 +12,23 @@ const server = createKoaServer({
   publicPath: '/assets',
   assetLocation: __dirname + '/assets',
 })
+
+if (process.env.USE_AUTH) {
+  // tslint:disable-next-line no-console
+  console.log(
+    `Protecting server using basic auth with user ${process.env.AUTH_NAME} 💂‍`,
+  )
+  const basicAuth = require('koa-basic-auth') // tslint:disable-line no-var-requires
+  const basicAuthMidleware = basicAuth({
+    name: notNullable(process.env.AUTH_NAME),
+    pass: notNullable(process.env.AUTH_PASS),
+  })
+  server.app.use(basicAuthMidleware)
+  server.router.use(basicAuthMidleware)
+} else {
+  // tslint:disable-next-line no-console
+  console.log('Not using any auth, server is open to the public')
+}
 
 reactPageRoutes.forEach((route) => {
   server.router.get(route.path, getPage)
