+++ src/utils/nullables.ts
@@ -1,0 +1,7 @@
+export const notNullable = <T>(thing?: T | null): T => {
+  if (thing === undefined || thing === null) {
+    throw new Error('Expected thing to be not nullable but was ' + typeof thing)
+  }
+
+  return thing
+}
